backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
pi==0.1.2
pipreqs==0.4.13
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarg==0.1.9
